requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==3.2.16
django-celery-beat==2.4.0
C==2.4.0
django-cors-headers==3.13.0
django-timezone-field==5.0
djangorestframework==3.14.0
dnspython==2.2.1
eventlet==0.33.2
greenlet==2.0.1
importlib-metadata==4.13.0
kombu==5.2.4
prompt-toolkit==3.0.36
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7
redis==4.4.0
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.7
vine==5.0.0
wcwidth==0.2.5
zipp==3.11.0
